+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/BugsnagTestUtils.java
@@ -2,8 +2,8 @@ package com.bugsnag.android;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
-import android.support.test.InstrumentationRegistry;
+import androidx.annotation.NonNull;
+import androidx.test.core.app.ApplicationProvider;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -46,7 +46,7 @@ final class BugsnagTestUtils {
 
     static Client generateClient(Configuration config) {
         config.setDelivery(generateDelivery());
-        return new Client(InstrumentationRegistry.getTargetContext(), config);
+        return new Client(ApplicationProvider.getApplicationContext(), config);
     }
 
     static Client generateClient() {
@@ -69,12 +69,13 @@ final class BugsnagTestUtils {
     }
 
     static Connectivity generateConnectivity() {
-        return new ConnectivityCompat(InstrumentationRegistry.getContext(), null);
+        return new ConnectivityCompat(ApplicationProvider.getApplicationContext(), null);
     }
 
     @NonNull
     static SessionStore generateSessionStore() {
-        return new SessionStore(generateConfiguration(), InstrumentationRegistry.getContext());
+        Context applicationContext = ApplicationProvider.getApplicationContext();
+        return new SessionStore(generateConfiguration(), applicationContext);
     }
 
     @SuppressWarnings("deprecation")