+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/ProjectPackagesTest.kt
@@ -1,6 +1,7 @@
 package com.bugsnag.android
 
-import android.support.test.InstrumentationRegistry
+import android.content.Context
+import androidx.test.core.app.ApplicationProvider
 import org.junit.Assert.assertArrayEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
@@ -12,7 +13,7 @@ class ProjectPackagesTest {
         val configuration = Configuration("api-key")
         assertNull(configuration.projectPackages)
 
-        val client = Client(InstrumentationRegistry.getContext(), configuration)
+        val client = Client(ApplicationProvider.getApplicationContext<Context>(), configuration)
         assertArrayEquals(arrayOf("com.bugsnag.android.core.test"), client.config.projectPackages)
         client.close()
     }
@@ -21,7 +22,7 @@ class ProjectPackagesTest {
     fun testProjectPackagesOverride() {
         val configuration = Configuration("api-key")
         configuration.projectPackages = arrayOf("com.foo.example")
-        val client = Client(InstrumentationRegistry.getContext(), configuration)
+        val client = Client(ApplicationProvider.getApplicationContext<Context>(), configuration)
         assertArrayEquals(arrayOf("com.foo.example"), client.config.projectPackages)
         client.close()
     }