+++ bugsnag-android-core/src/main/java/com/bugsnag/android/MapUtils.java
@@ -1,6 +1,6 @@
 package com.bugsnag.android;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.Map;