+++ features/fixtures/mazerunner/src/main/java/com/bugsnag/android/mazerunner/scenarios/CXXImproperTypecastScenario.java
@@ -5,7 +5,7 @@ import android.content.Context;
 
 import com.bugsnag.android.Configuration;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class CXXImproperTypecastScenario extends Scenario {